Moved some types to a separate file
