+++ tests/unittest/public/test_managebot_autojoin.py
@@ -5,8 +5,8 @@ from io import StringIO
 
 from asynctest.mock import Mock, patch
 
-from source.data import Message
 from source.database import AutoJoinChannel, DatabaseMain
+from source.data.message import Message
 from tests.unittest.base_managebot import TestManageBot, send
 from tests.unittest.mock_class import AsyncIterator, StrContains
